+++ src/app/home/initial-page/components/initial-page.component.ts
