+++ src/app/home/initial-page/initial-page.module.ts
@@ -1,11 +1,11 @@
 import { TranslateModule } from '@ngx-translate/core';
 import { NgModule } from '@angular/core';
-import { AppRoutingModule } from '../app-routing.module';
+import { AppRoutingModule } from '../../app-routing.module';
 import { CommonModule } from '@angular/common';
-import { CoreModule } from '../core/core.module';
-import { LayoutModule } from '../layout/layout.module';
+import { CoreModule } from '../../core/core.module';
+import { LayoutModule } from '../../layout/layout.module';
 
-import { InitialPageComponent } from './initial-page/initial-page.component';
+import { InitialPageComponent } from './components/initial-page.component';
 
 /* @export
  * @class HomeModule
@@ -22,4 +22,4 @@ import { InitialPageComponent } from './initial-page/initial-page.component';
   declarations: [InitialPageComponent],
   exports: [InitialPageComponent],
 })
-export class HomeModule {}
+export class InitialPageModule {}
