+++ src/app/home/sampledata/sampledata-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AuthGuard } from '../../core/security/auth-guard.service';
 import { NavBarComponent } from '../../layout/nav-bar/nav-bar.component';
-import { InitialPageComponent } from '../initial-page/initial-page.component';
+import { InitialPageComponent } from '../initial-page/components/initial-page.component';
 import { SampleDataGridComponent } from './components/sampledata-grid/sampledata-grid.component';
 
 const routes: Routes = [
