+++ src/app/shared/models/search-criteria.ts
